Pause / Stop
Changed so it outputs „Paused“ when the player is paused, and „Stopped“
when the player is stopped. I had some issues with OBS when the .txt
file was empty (lots of errors in the logfile, which got bigger and
bigger, up to 60 MB). Now when the script isn’t running the .txt file
always contains „Stopped“.